move xarray utils to own file
